requirements.txt
altgraph==0.17.4
asttokens==2.4.1
auto-py-to-exe==2.42.0
bokeh==3.3.2
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Eel==0.16.0
executing==2.0.1
fonttools==4.47.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pyqtgraph==0.13.3
pyserial==3.5
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
qt6-applications==6.5.0.2.3
qt6-tools==6.5.0.1.3
scipy==1.11.4
shiboken6==6.6.0
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.4
wcwidth==0.2.10
whichcraft==0.6.1
xyzservices==2023.10.1
zope.event==5.0
zope.interface==6.1
